backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0
# asyncpg==0.29.0  # PostgreSQL (Phase 2)

# HTTP 客户端
httpx==0.26.0
requests==2.31.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSocket
websockets==12.0

# 工具库
fake-useragent==1.4.0
python-dotenv==1.0.0

# 测试 (Phase 1)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Celery + Redis (Phase 2)
# celery==5.3.6
# redis==5.0.1

# 生产部署 (Phase 2)
gunicorn==21.2.0

# 日志
# loguru==0.7.2  # 可选：更好的日志库
